requirements.txt
# Core framework
fastapi==0.115.0
uvicorn[standard]==0.30.6

# MongoDB async driver
motor==3.6.0

# Auth (JWT + password hashing)
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# AWS SDK
boto3==1.35.21
botocore==1.35.21

# WebSockets
websockets==12.0

# Config
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.4.0
pydantic[email]

# Utils
structlog==24.4.0
tenacity==9.0.0

# Dev/test
httpx==0.27.2
pytest==8.3.2

slowapi==0.1.9          # rate limiting (Starlette/FastAPI)
user-agents==2.2.0      # optional: user-agent parsing for logging/rules
